+++ packages/instant-apps-components/src/components/instant-apps-export/instant-apps-export.tsx
@@ -558,6 +558,7 @@ export class InstantAppsExport {
         this.legend = new Legend({
           container: legendCont,
           view,
+          respectLayerDefinitionExpression: true,
           style: {
             type: 'card',
             layout: 'side-by-side',