+++ packages/instant-apps-components/src/components/instant-apps-measurement/instant-apps-measurement.tsx
@@ -69,8 +69,8 @@ export class InstantAppsMeasurement {
     return (
       <Host>
         <calcite-panel heading={this?.messages?.label} closable={this.closable}>
-          <div>{this.renderActionPad()}</div>
-          {this.renderMeasureToolPanel()}
+          <div slot="content-top">{this.renderActionPad()}</div>
+          <div>{this.renderMeasureToolPanel()}</div>
         </calcite-panel>
       </Host>
     );
@@ -102,7 +102,7 @@ export class InstantAppsMeasurement {
           data-value="distance"
           onClick={this._handleToolClick.bind(this)}
         >
-          <calcite-tooltip close-on-click={true} placement="bottom" slot="tooltip">
+          <calcite-tooltip close-on-click={true} placement="top" slot="tooltip">
             {messages?.line}
           </calcite-tooltip>
         </calcite-action>
@@ -114,7 +114,7 @@ export class InstantAppsMeasurement {
           data-value="area"
           onClick={this._handleToolClick.bind(this)}
         >
-          <calcite-tooltip close-on-click={true} placement="bottom" slot="tooltip">
+          <calcite-tooltip close-on-click={true} placement="top" slot="tooltip">
             {messages?.area}
           </calcite-tooltip>
         </calcite-action>
@@ -126,13 +126,13 @@ export class InstantAppsMeasurement {
           data-value="point"
           onClick={this._handleToolClick.bind(this)}
         >
-          <calcite-tooltip close-on-click={true} placement="bottom" slot="tooltip">
+          <calcite-tooltip close-on-click={true} placement="top" slot="tooltip">
             {messages?.point}
           </calcite-tooltip>
         </calcite-action>
 
         <calcite-action text={messages?.clear} scale="m" icon="trash" data-value="clear" onClick={this._handleToolClick.bind(this)}>
-          <calcite-tooltip close-on-click={true} placement="bottom" slot="tooltip">
+          <calcite-tooltip close-on-click={true} placement="top" slot="tooltip">
             {messages?.clear}
           </calcite-tooltip>
         </calcite-action>