+++ packages/instant-apps-components/CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.0.0-beta.269
+
+### instant-apps-sign-in
+
+- add `user-id` to calcite-avatar
+
+### instant-apps-export
+
+- add `respectLayerDefinitionExpression` to Legend
+
+## v1.0.0-beta.267
+
+### instant-apps-language-switcher
+
+- update `selectedLanguage` to a prop
+
 ## v1.0.0-beta.266
 
 ### instant-apps-app-guide